src/third_party/blink/web_tests/accessibility/name-calc-svg.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/run-after-layout-and-paint.js"></script>

<div class="container">
  <svg id="svg1">
    <title>svg1-title</title>
  </svg>
</div>

<script>
test_after_layout_and_paint(function(t) {
  var axSvg1 = accessibilityController.accessibleElementById("svg1");
  assert_equals(axSvg1.name, "svg1-title");
  assert_equals(axSvg1.nameFrom, "relatedElement");
}, "SVG element with title element");
</script>
